Fix ordering issue that prevented including the ::kafo class
The Exec that ran librarian-puppet was in the same dynamically generated Puppet
manifest as the call to ::kafo class. This meant that puppet apply did not have
access to the ::kafo class when it loaded the manifest and hence refused to
apply anything, including running the librarian-puppet Exec. Fix by moving the
Exec to first Puppet run cycle, so that when the second launches the modulepath
and ::kafo is ensured to be available.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>