Remove parameter hosts_entries from all classes. This handling of hashes is a kafo limitation.
See https://community.theforeman.org/t/kafo-wizard-input-a-hash/11049

The ip's and hosts entries
belong to DNS as an A record anyway. Or the user needs to enter
the ip's and the entries manually to the hosts file.

Bring it back if/when we need it.
If we find it will be needed, work around the kafo limitation with
a proper ruby function.

Signed-off-by: Petri Lammi <petri.lammi@puppeteers.fi>