Add mappings to all scenarios
The param(module,param) method in hook_context.rb in Kafo does not work if the
module is not a top-level class. In other words, 'puppetmaster::puppetboard'
does not work, whereas 'puppetmaster_puppetboard' (a mapping) does. And we need
the param method to work if we want to dynamically configure some parameters
(answers) if they're missing. The immediate goal is to automatically create
random passwords if the user does not specify them, but the params method can be
used for other things as well.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>