Automatically set PuppetDB password to a random string if unset
This improves the usability of the installer, because it is highly unlikely that
all users would be disciplined enough to go through installer parameters before
selecting "Save and run". Also, the Ruby backtrace for unset, mandatory
parameters is really ugly and user-unfriendly, so having sane (random) defaults
for passwords makes lots of sense.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>