Copy eyaml keys into the correct place
The eyaml keysshould go to /etc/puppetlabs/puppet/eyaml/keys,
but they were copied to /etc/puppetlabs/puppet/eyaml. This left the installation
with two pairs of eyaml keys if user provides his/her own: the wrong (generated)
ones in the correct directory, and the correct ones in the wrong directory. Fix
this.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>