Kafo: add pre hook for sanity checking control repo settings
The code around control repository automation is quite complex, so the user
experience used not to be that great. Do full validation of parameter
interactions/dependencies and also test if the r10k_key is present before
attempting to run the installer. This only helps with the interactive installer,
though.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>